Autodetect PWADs and smarter IWAD search
The header of the WAD will be read to see if it's a IWAD/PWAD
and load it with -iwad/-file respectively.
Before this change it was not possible to load PWADs properly.

Also, improved the lookup for WADs by checking both in the wad folder
and system folder. If the WAD is still not found it will look in the
parent folders recursively.