Emulate the mouse with the right analog stick
This gives analog camera movement when using the right stick by
outputing PrBoom mouse movement with it.

Also because it does not depend on keybindings, the right stick
function won't get messed up when the rotation controls are changed
to something else than "rarrow" and "larrow" in the in-game options.

The threshold value is still used for proportionality in the deadzone
and multiplier.