Tweak mouse sensitivity default & analog mouse
The default mouse sensitivity was very low (10/100), which made it so
for any sane use of the mouse to control the camera you'd need to
increase the sensitivity considerably.

But increasing the sensitivity of the mouse now affects the sensitivity
of the analog too, so this was causing problems, since the analog was
set to work fine with default settings.

This change tweaks the value for the analog mouse movement to make it
reasonably lower while increasing the mouse sensitivity default.
This way it's possible to have the same prboom.cfg values be suitable
for both mouse and gamepad, and the default installation of the core
will have a sane default when using it with a mouse as well.