I_FindFile: always use the extension when provided
Previous find function was checking first for files without matching the
extension and only if no matches were found tried appending the extension.

This can cause problems when there are files with no extension matching
the deh/iwad filename or, in linux, a directory with the same name as
the file.

There's no apparent reason why the old prboom code might have been doing
this, probably for convenient when calling it from the terminal without
specifying the full filename, but this wouldn't apply to libretro, since
we require full filename.