Adjust game speed in relation to framerate
The game was previously getting  accelerated when the Framerate option
was set to anything higher than 35fps. This fixes that.

The display will still be refreshed at a higher rate, but the
interpolation code appears to have been scrapped when the engine was
ported to libretro so it still needs to be re-added. But this is would
be the first step to make higher framerate work.

Also, I noticed that setting the Framerate option to "60fps" was making
it run at 120fps instead. So I've changed it to 60 properly and added a
new 120fps option, since the "retro_get_system_av_info" of the core is
actually set to run at 120, and some frontends (like Kodi) expect it to
match the core's FPS.