Keybinding-independent gamepad & input changes
- The gamepad bindings now will work independently of what the user has
configured in PrBoomm keybinding settings. Finally it'll be possible to
set a WASD configuration for keyboard without messing everything up
whenever you feel like plugging a gamepad in the same setup.

- Removed setting for changing Menu buttons based on their position,
since it was conflicting with Retroarch's remappings. Instead a fixed
set of game and menu actions was associated to each gamepad layout.
There are some redundancies in the menu actions so there's some level
of flexibility if the user wants to remap the buttons.

- The modern gamepad is now modeled after the controller mappings from
the original XBOX Doom 3 Collection version of classic Doom. While it's
unfortunate that the bindings are changed (again) I'm hoping this will
the last time, it was done with some consensus and the change was needed
to avoid confusing conflicts with the menu actions switching when
switching between classic<->modern

- The mouse wheel now also scrolls through the options of the menu

- Some changes in the left analog to ease it when on the menu.