Load default config base values from scanned files
With this change, the default values for the PrBoom configuration
options will be loaded from prboom.cfg files found in the WAD folder
or in the system folder (whatever is found first).

If a prboom.cfg file exists in the same folder of the WAD that is prefixed
with the WAD name and a dot (eg. plutonia.prboom.cfg for plutonia.wad) then
that cfg file will be the one loaded as base default for that WAD.

These are only affecting the base default values, the user configured values
would still be kept in the prboom.cfg file inside the save folder of
each WAD.

In order to improve the reusability of the config files, when saving the
user configuration, the default values will be saved as comments and only
the values that differ from the base default will be uncommented.