Use gamepad modern D-pad to move around map
Before this, if the user sets WASD keys for movement there was no way to
scroll around in the automap view.
Those who did not change the default keys will not be affected by this
change (because the default map movement uses the same arrow keys as the
movement keys), but those who do change it will be able to scroll the
automap now.