Make prboom.wad optional, not required
prboom.wad will still be loaded and used if present, but will be no
longer a requirement for the core to run.

The PrBoom-specific graphics from the prboom.wad fallback to some
graphics from original Doom IWADs when not found.

Several of the required tables are now loaded in the code itself, and
will be used as the default when the corresponding lumps are not found
(so it is still possible for mods to override them).

This is similar as to how the original Doom (and chocolate doom) work,
which also do not require any additional data other than that on the
WAD itself.

TODO: WATERMAP fallback and improving font fallbacks
Note that TNT1A0 has no fallback but it'll still not be drawn
and HUD graphics are at the moment not really being used.