New color translation for option menu highlight
Currently the CRORANGE colormap is used to highlight entries in the options.
This color translation table only changes some red colors into orange, so
it only works when the text is red, such as in the original Doom iwads.

However, many PWADs override the font to use different colors, which makes
the options menu very uncomfortable to use since you cannot tell what entry
is selected.

A color translation table was added which changes most of the palette
shifting it to the lighter colors, in order to have a visible change
even if the font is replaced.