Warning fixes/supressions
With this there are no more warnings when compiling with mingw and with
clang from android-ndk.

There are some remaining implicit declaration and int conversion
warnings in libretro-common when compiling using latest gcc on linux.
But maybe they should be first fixed first in the upstream
libretro-common so they are not reintroduced later on.