Allow using the Heads UP Display
The options hud_display and hud_distributed were merged into a single
"hud_mode" option.

The position of health/armor and stats/keys in the distributed layout
were switched.

The stats line was optimized for screen space, removing the static "STS"
text as well as some of the spacing, and the items collected+total was
combined in an items left count.

The keys line won't be shown if there are no keys.