Play PrBoom+/UMAPINFO demos flagged
PrBoom+/UMAPINFO port prepends a 255 byte to the version of the demo
whenever UMAPINFO is being used, so that incompatible engines will
fail to detect the version.

This adds support for it allowing the rest of the versioning system
to be kept as it is and let UMAPINFO be used with all existing
formats. In case of a mismatch an error will be printed but the demo
be allowed to play regardless of the mismatch.